requirements.txt
Cython==0.29.20
matplotlib==3.2.2
numpy==1.18.5
pandas==1.0.5
pickleshare==0.7.5
Pillow==7.1.2
scikit-learn==0.23.2
scipy==1.5.0
seaborn==0.10.1
torch==1.5.1
torchvision==0.6.0
tqdm==4.46.1
urllib3==1.25.9
